Reorganize JHLite steps
