+++ src/test/java/tech/jhipster/lite/TestProjects.java
@@ -1,0 +1,16 @@
+package tech.jhipster.lite;
+
+public final class TestProjects {
+
+  private static String lastProjectFolder;
+
+  private TestProjects() {}
+
+  public static String newTestFolder() {
+    return lastProjectFolder = TestFileUtils.tmpDirForTest();
+  }
+
+  public static String lastProjectFolder() {
+    return lastProjectFolder;
+  }
+}
